% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citius-package.R
\docType{package}
\name{citius-package}
\alias{citius}
\alias{citius-package}
\title{citius: Performance Modelling and Outcome Simulation for Major Games Athletics and Swimming}
\description{
Estimates latent athlete ability from competition result histories and simulates event outcomes for major multi-sport games (Olympic, Commonwealth, World Championships). Combines decay-weighted empirical-Bayes ability estimation with event-specific performance variance, aging curves, shared race-condition effects and round progression (heats, semi-finals, finals) in a Monte Carlo engine that returns finishing-position, medal and mark-threshold probabilities. Data is sourced from World Athletics and World Aquatics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/peteowen1/citius}
  \item \url{https://peteowen1.github.io/citius/}
  \item Report bugs at \url{https://github.com/peteowen1/citius/issues}
}

}
\author{
\strong{Maintainer}: Pete Owen \email{pete.owen1@hotmail.co.uk} [copyright holder]

}
\keyword{internal}
