% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema.R
\docType{data}
\name{CITIUS_DB_SCHEMA}
\alias{CITIUS_DB_SCHEMA}
\title{Expected columns per citius DuckDB store table}
\format{
An object of class \code{list} of length 3.
}
\usage{
CITIUS_DB_SCHEMA
}
\description{
Expected columns per citius DuckDB store table
}
\keyword{internal}
