% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_store.R
\name{citius_table_exists}
\alias{citius_table_exists}
\title{Does a table exist in the citius DuckDB database?}
\usage{
citius_table_exists(conn, table_name)
}
\arguments{
\item{conn}{Connection.}

\item{table_name}{Unqualified table name.}
}
\value{
Logical.
}
\description{
Does a table exist in the citius DuckDB database?
}
\keyword{internal}
