% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_store.R
\name{.check_citius_duckdb_available}
\alias{.check_citius_duckdb_available}
\title{Check DuckDB/DBI are usable}
\usage{
.check_citius_duckdb_available()
}
\description{
Check DuckDB/DBI are usable
}
\keyword{internal}
